refactor: consolidate WorkflowStep/WorkflowDefinition types into shared module (#735)
